models

+++ RestApi3/RestApi3/src/Models/Product.cs
@@ -5,7 +5,7 @@ public class Product
     public int IdProduct { get; init; }
     public string Name { get; init; }
     public string Description { get; init; }
-    public int Price { get; init; }
+    public double Price { get; init; }
 
     public Product(int idProduct, string name, string description, int price)
     {
